requirements.txt
streamlit==1.28.0
langchain==0.0.346
openai==1.3.0
chromadb==0.4.15
python-docx==0.8.11
PyPDF2==3.0.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.17.0